+++ src-java/flowmonitoring-topology/flowmonitoring-storm-topology/src/main/java/org/openkilda/wfm/topology/flowmonitoring/bolt/FlowSplitterBolt.java
@@ -15,13 +15,18 @@
 
 package org.openkilda.wfm.topology.flowmonitoring.bolt;
 
+import static org.openkilda.wfm.topology.flowmonitoring.FlowMonitoringTopology.Stream.FLOW_REMOVE_STREAM_ID;
 import static org.openkilda.wfm.topology.flowmonitoring.FlowMonitoringTopology.Stream.FLOW_UPDATE_STREAM_ID;
 import static org.openkilda.wfm.topology.flowmonitoring.bolt.FlowCacheBolt.FLOW_ID_FIELD;
 import static org.openkilda.wfm.topology.utils.KafkaRecordTranslator.FIELD_ID_PAYLOAD;
 
+import org.openkilda.messaging.Message;
+import org.openkilda.messaging.command.CommandData;
+import org.openkilda.messaging.command.CommandMessage;
 import org.openkilda.messaging.info.InfoData;
 import org.openkilda.messaging.info.InfoMessage;
-import org.openkilda.messaging.info.flow.UpdateFlowInfo;
+import org.openkilda.messaging.info.flow.RemoveFlowCommand;
+import org.openkilda.messaging.info.flow.UpdateFlowCommand;
 import org.openkilda.messaging.info.stats.FlowRttStatsData;
 import org.openkilda.wfm.AbstractBolt;
 import org.openkilda.wfm.error.PipelineException;
@@ -34,17 +39,35 @@ import org.apache.storm.tuple.Values;
 public class FlowSplitterBolt extends AbstractBolt {
 
     public static final String INFO_DATA_FIELD = "info-data-field";
+    public static final String COMMAND_DATA_FIELD = "command-data-field";
 
     @Override
     protected void handleInput(Tuple input) throws PipelineException {
-        InfoData payload = pullValue(input, FIELD_ID_PAYLOAD, InfoMessage.class).getData();
-        if (payload instanceof FlowRttStatsData) {
-            FlowRttStatsData flowRttStatsData = (FlowRttStatsData) payload;
-            emit(input, new Values(flowRttStatsData.getFlowId(), flowRttStatsData, getCommandContext()));
-        } else if (payload instanceof UpdateFlowInfo) {
-            UpdateFlowInfo updateFlowInfo = (UpdateFlowInfo) payload;
-            emit(FLOW_UPDATE_STREAM_ID.name(), input, new Values(updateFlowInfo.getFlowId(), updateFlowInfo,
-                    getCommandContext()));
+        Message message = pullValue(input, FIELD_ID_PAYLOAD, Message.class);
+        if (message instanceof InfoMessage) {
+            InfoData infoData = ((InfoMessage) message).getData();
+            if (infoData instanceof FlowRttStatsData) {
+                FlowRttStatsData flowRttStatsData = (FlowRttStatsData) infoData;
+                emit(input, new Values(flowRttStatsData.getFlowId(), flowRttStatsData, getCommandContext()));
+            } else {
+                unhandledInput(input);
+            }
+            return;
+        }
+
+        if (message instanceof CommandMessage) {
+            CommandData commandData = pullValue(input, FIELD_ID_PAYLOAD, CommandMessage.class).getData();
+            if (commandData instanceof UpdateFlowCommand) {
+                UpdateFlowCommand updateFlowCommand = (UpdateFlowCommand) commandData;
+                emit(FLOW_UPDATE_STREAM_ID.name(), input, new Values(updateFlowCommand.getFlowId(), updateFlowCommand,
+                        getCommandContext()));
+            } else if (commandData instanceof RemoveFlowCommand) {
+                RemoveFlowCommand removeFlowCommand = (RemoveFlowCommand) commandData;
+                emit(FLOW_REMOVE_STREAM_ID.name(), input, new Values(removeFlowCommand.getFlowId(),
+                        getCommandContext()));
+            } else {
+                unhandledInput(input);
+            }
         } else {
             unhandledInput(input);
         }
@@ -54,6 +77,7 @@ public class FlowSplitterBolt extends AbstractBolt {
     public void declareOutputFields(OutputFieldsDeclarer declarer) {
         declarer.declare(new Fields(FLOW_ID_FIELD, INFO_DATA_FIELD, FIELD_ID_CONTEXT));
         declarer.declareStream(FLOW_UPDATE_STREAM_ID.name(),
-                new Fields(FLOW_ID_FIELD, INFO_DATA_FIELD, FIELD_ID_CONTEXT));
+                new Fields(FLOW_ID_FIELD, COMMAND_DATA_FIELD, FIELD_ID_CONTEXT));
+        declarer.declareStream(FLOW_REMOVE_STREAM_ID.name(), new Fields(FLOW_ID_FIELD, FIELD_ID_CONTEXT));
     }
 }