+++ src-java/kilda-persistence-api/src/main/java/org/openkilda/persistence/repositories/RepositoryFactory.java
@@ -58,6 +58,8 @@ public interface RepositoryFactory {
 
     SwitchPropertiesRepository createSwitchPropertiesRepository();
 
+    FlowStatsRepository createFlowStatsRepository();
+
     SwitchConnectedDeviceRepository createSwitchConnectedDeviceRepository();
 
     PortEventRepository createPortEventRepository();