+++ src-java/testing/functional-tests/src/test/groovy/org/openkilda/functionaltests/spec/flows/FlowCrudV1Spec.groovy
@@ -874,7 +874,7 @@ class FlowCrudV1Spec extends HealthCheckSpecification {
                     new InstallIngressFlow(UUID.randomUUID(), NON_EXISTENT_FLOW_ID, null, sw.dpId,
                             5, 6, 5, 0, meterId, FlowEncapsulationType.TRANSIT_VLAN,
                             OutputVlanType.REPLACE, fakeBandwidth, meterId, sw.dpId, false, false, false, null))
-                    .toJson()))
+                    .toJson())).get()
         }
         producer.close()